+++ lib/Extension/LanguageServerCodeTransform/CodeAction/GenerateDecoratorProvider.php
@@ -4,6 +4,7 @@ namespace Phpactor\Extension\LanguageServerCodeTransform\CodeAction;
 
 use Amp\CancellationToken;
 use Amp\Promise;
+use Phpactor\Extension\LanguageServerBridge\Converter\TextDocumentConverter;
 use Phpactor\Extension\LanguageServerCodeTransform\LspCommand\GenerateDecoratorCommand;
 use Phpactor\LanguageServerProtocol\CodeAction;
 use Phpactor\LanguageServerProtocol\Command;
@@ -32,13 +33,12 @@ class GenerateDecoratorProvider implements CodeActionProvider
     public function provideActionsFor(TextDocumentItem $textDocument, Range $range, CancellationToken $cancel): Promise
     {
         return call(function () use ($textDocument) {
-            $classes = $this->reflector->reflectClassesIn($textDocument->text);
-            if (count($classes) !== 1) {
+            $classes = $this->reflector->reflectClassesIn(TextDocumentConverter::fromLspTextItem($textDocument));
+            if (count($classes->classes()) !== 1) {
                 return [];
             }
 
-
-            $class = $classes->first();
+            $class = $classes->classes()->first();
 
             if (!$class instanceof ReflectionClass) {
                 return [];