+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/DeprecatedUsageDiagnostic.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Phpactor\WorseReflection\Bridge\TolerantParser\Diagnostics;
+
+use Phpactor\TextDocument\ByteOffsetRange;
+use Phpactor\WorseReflection\Core\Diagnostic;
+use Phpactor\WorseReflection\Core\DiagnosticSeverity;
+
+class DeprecatedUsageDiagnostic implements Diagnostic
+{
+    public function __construct(private ByteOffsetRange $range, private string $memberName, private string $message, private string $memberType)
+    {
+    }
+
+    public function range(): ByteOffsetRange
+    {
+        return $this->range;
+    }
+
+    public function severity(): DiagnosticSeverity
+    {
+        return DiagnosticSeverity::WARNING();
+    }
+
+    public function message(): string
+    {
+        if (!$this->message) {
+            return sprintf('Call to deprecated %s "%s"', $this->memberType, $this->memberName);
+        }
+
+        return sprintf('Call to deprecated %s "%s": %s', $this->memberType, $this->memberName, $this->message);
+    }
+}