+++ lib/WorseReflection/Bridge/TolerantParser/Reflection/ReflectionClass.php
@@ -361,6 +361,11 @@ class ReflectionClass extends AbstractReflectionClass implements CoreReflectionC
         return ReflectionClassLikeCollection::fromReflections((new ClassHierarchyResolver())->resolve($this));
     }
 
+    public function classLikeType(): string
+    {
+        return 'class';
+    }
+
     protected function node(): Node
     {
         return $this->node;