+++ lib/WorseReflection/Bridge/TolerantParser/Reflection/ReflectionEnum.php
@@ -111,6 +111,11 @@ class ReflectionEnum extends AbstractReflectionClass implements CoreReflectionEn
         return NodeUtil::typeFromQualfiedNameLike($this->serviceLocator()->reflector(), $this->node, $this->node->enumType);
     }
 
+    public function classLikeType(): string
+    {
+        return 'enum';
+    }
+
     /**
      * @return EnumDeclaration
      */