+++ lib/WorseReflection/Core/Inference/Context/FunctionCallContext.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference\Context;
+
+use Phpactor\TextDocument\ByteOffsetRange;
+use Phpactor\WorseReflection\Core\Inference\NodeContext;
+use Phpactor\WorseReflection\Core\Inference\Symbol;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionFunction;
+
+class FunctionCallContext extends NodeContext
+{
+    public function __construct(Symbol $symbol, private ByteOffsetRange $byteOffsetRange, private ReflectionFunction $function)
+    {
+        parent::__construct($symbol, $function->inferredType()->reduce());
+    }
+
+    public function range(): ByteOffsetRange
+    {
+        return $this->byteOffsetRange;
+    }
+
+    public function function(): ReflectionFunction
+    {
+        return $this->function;
+    }
+}