+++ lib/WorseReflection/Core/Inference/Context/PropertyAccessContext.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Phpactor\WorseReflection\Core\Inference\Context;
+
+use Phpactor\WorseReflection\Core\Reflection\ReflectionProperty;
+
+/**
+ * @extends MemberAccessContext<ReflectionProperty>
+ */
+class PropertyAccessContext extends MemberAccessContext
+{
+}