+++ lib/WorseReflection/Core/Inference/Resolver/MemberAccess/NodeContextFromMemberAccess.php
@@ -7,7 +7,9 @@ use Microsoft\PhpParser\Node\DelimitedList\TraitSelectOrAliasClauseList;
 use Microsoft\PhpParser\Node\Expression\CallExpression;
 use Microsoft\PhpParser\Node\Expression\MemberAccessExpression;
 use Microsoft\PhpParser\Node\Expression\ScopedPropertyAccessExpression;
+use Phpactor\TextDocument\ByteOffsetRange;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
+use Phpactor\WorseReflection\Core\Inference\Context\MemberAccessContext;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\FunctionArguments;
 use Phpactor\WorseReflection\Core\Inference\GenericMapResolver;
@@ -55,7 +57,11 @@ class NodeContextFromMemberAccess
         $memberTypeName = $node->getParent() instanceof CallExpression ? Symbol::METHOD : Symbol::PROPERTY;
 
         // support trait method-alias clauses, e.g. use  use A, B {A::foobar insteadof B; B::bigTalk insteadof A;}
-        if ($memberTypeName === Symbol::PROPERTY && $node->parent->parent && $node->parent->parent instanceof TraitSelectOrAliasClauseList) {
+        if (
+            $memberTypeName === Symbol::PROPERTY &&
+            $node->parent?->parent &&
+            $node->parent->parent instanceof TraitSelectOrAliasClauseList
+        ) {
             $memberTypeName = Symbol::METHOD;
         }
 
@@ -93,7 +99,7 @@ class NodeContextFromMemberAccess
             }
         }
 
-        [ $containerType, $memberType ] = $this->resolveContainerMemberType(
+        [ $containerType, $memberType, $member ] = $this->resolveContainerMemberType(
             $resolver,
             $frame,
             $node,
@@ -106,13 +112,23 @@ class NodeContextFromMemberAccess
             $containerType = $classType;
         }
 
+        if ($member instanceof ReflectionMember) {
+            return new MemberAccessContext(
+                $context->symbol(),
+                $memberType->reduce(),
+                $containerType,
+                ByteOffsetRange::fromInts($node->memberName->getStartPosition(), $node->memberName->getEndPosition()),
+                $member,
+            );
+        }
+
         return $context->withContainerType(
             $containerType
         )->withType($memberType->reduce());
     }
 
     /**
-     * @return array{Type,Type}
+     * @return array{Type,Type,?ReflectionMember}
      */
     private function resolveContainerMemberType(
         NodeContextResolver $resolver,
@@ -124,6 +140,7 @@ class NodeContextFromMemberAccess
     ): array {
         $types = [];
         $memberType = TypeFactory::undefined();
+        $member = null;
 
         $arguments = $this->resolveArguments($resolver, $frame, $node->parent);
         // this could be a union or a nullable
@@ -158,7 +175,7 @@ class NodeContextFromMemberAccess
         }
 
         $containerType = UnionType::fromTypes(...$types)->reduce();
-        return [$containerType, $memberType];
+        return [$containerType, $memberType, $member];
     }
 
     private function resolveMemberType(NodeContextResolver $resolver, Frame $frame, ReflectionMember $member, ?FunctionArguments $arguments, Node $node, Type $subType): Type
@@ -237,9 +254,9 @@ class NodeContextFromMemberAccess
         }
 
         $variable = $frame->properties()
-        ->lessThanOrEqualTo($position)
-        ->byName($propertyName)
-        ->lastOrNull();
+            ->lessThanOrEqualTo($position)
+            ->byName($propertyName)
+            ->lastOrNull();
 
         if (null === $variable) {
             return null;